Address review feedback: refactored save.cpp and flash.cpp
