+++ src/main.cpp
@@ -1,7 +1,9 @@
 #include <Arduino.h>
 #include <blink.h>
 #include <tmp.h>
-#include "flash.h"  // SPI flash support
+#include <flash.h>  // SPI flash support
+#include <save.h>   // saveState functionality
+
 
 void setup() {
     // -------------------- Setup --------------------
